Add authentication library update functionality to index.php
Introduced functionality to retrieve and update the 'mardens-auth-lib.js' authentication library from 'auth.mardens.com' in 'index.php'. The update is either performed if the file doesn't exist or if a newer version is available. SSL errors are ignored and HTTP GET requests are used to interact with the server.